#1.0.1 给二维码中间添加logo 修改了我numbertool精度问题，float精度只有6位替换成double

+++ LaxoWallet/LWModule/Home/View/new/LWHomeListHeadView.m
@@ -134,7 +134,7 @@
         return;
     }
     
-    CGFloat bitCount = 0;
+    NSInteger bitCount = 0;
     for (NSInteger i = 0; i<currentArray.count; i++) {
         LWHomeWalletModel *dataModel = [currentArray objectAtIndex:i];
         for (NSInteger j = 0; j<dataModel.utxo.count; j++) {
